TetraNyte Simultaneous Multithreaded conformance tests
