--- rtl/TetraNyte/rv32i_Zmmul/src/main/scala/TetraNyteRV32IZmmulCoreMRF.scala
@@ -1,410 +1,0 @@
-package TetraNyte
-
-import chisel3._
-import chisel3.util._
-import chisel3.dontTouch
-import ALUs._
-import BranchUnit._
-import Decoders._
-import LoadUnit._
-import StoreUnit._
-import RegFiles._
-
-class TetraNyteRV32IZmmulCoreMRF extends Module {
-  val numThreads = 4
-  val io = IO(new TetraNyteRV32IZmmulCoreIO(numThreads))
-
-  // Per-thread PC registers and flush tracking
-  val pcResetVec = VecInit(Seq.fill(numThreads)("h80000000".U(32.W)))
-  val flushResetVec = VecInit(Seq.fill(numThreads)(false.B))
-  val pcRegs = Reg(Vec(numThreads, UInt(32.W)))
-  val flushThread = RegInit(VecInit(Seq.fill(numThreads)(false.B)))
-
-  // Long-latency division bookkeeping (global stall while active to keep hazards simple)
-  val divider = Module(new Div32Radix4)
-  val divActive = RegInit(false.B)
-  val divThread = Reg(UInt(log2Ceil(numThreads).W))
-  val divRd = Reg(UInt(5.W))
-  val divFunct3 = Reg(UInt(3.W))
-  val divDividend = Reg(UInt(32.W))
-  val divDivisor = Reg(UInt(32.W))
-  val divResult = Reg(UInt(32.W))
-  val divDoneSticky = RegInit(false.B)
-  val divLaunch = WireDefault(false.B)
-  // No global stall: allow other threads to proceed while a divide runs.
-
-  // Round-robin thread scheduler
-  val threadSel = RegInit(0.U(log2Ceil(numThreads).W))
-  io.fetchThread := threadSel
-
-  // Single pipeline registers (carry threadId with each instruction)
-  // 4-stage pipeline: IF, ID, EX, MEM+WB (combined)
-  val if_id = RegInit(0.U.asTypeOf(new PipelineRegBundle))
-  val id_ex = RegInit(0.U.asTypeOf(new PipelineRegBundle))
-  val ex_mem = RegInit(0.U.asTypeOf(new PipelineRegBundle))
-
-  // Shared multithreaded register file
-  val regFile = Module(new RegFileMT2R1WMem(numThreads = numThreads))
-  val unusedRegDebugX1 = Wire(Vec(numThreads, UInt(32.W)))
-  unusedRegDebugX1 := regFile.io.debugX1
-  dontTouch(unusedRegDebugX1)
-
-  // Debug mirrors to expose last-seen per-thread stage values
-  val debugIfInstr = RegInit(VecInit(Seq.fill(numThreads)(0.U(32.W))))
-  val debugIdRs1 = RegInit(VecInit(Seq.fill(numThreads)(0.U(32.W))))
-  val debugIdRs2 = RegInit(VecInit(Seq.fill(numThreads)(0.U(32.W))))
-  val debugExAlu = RegInit(VecInit(Seq.fill(numThreads)(0.U(32.W))))
-  val debugMemLoad = RegInit(VecInit(Seq.fill(numThreads)(0.U(32.W))))
-  val memLoadData = WireDefault(0.U(32.W))
-
-  // Default IO outputs
-  io.memAddr := 0.U
-  io.memWrite := 0.U
-  io.memMask := 0.U
-  io.memValid := false.B
-  io.memMisaligned := false.B
-
-  // ===================== Instruction Fetch (IF) =====================
-  val currentThread = threadSel
-  val flushThisThread = flushThread(currentThread)
-  val threadEnabled = io.threadEnable(currentThread)
-  val currentPC = pcRegs(currentThread)
-
-  // IF: always update the IF/ID register for the currently selected thread.
-  if_id.threadId := currentThread
-  if_id.pc := currentPC
-  if_id.instr := io.instrMem
-  if_id.rs1 := io.instrMem(19, 15)
-  if_id.rs2 := io.instrMem(24, 20)
-  if_id.rd := io.instrMem(11, 7)
-  if_id.valid := !flushThisThread && threadEnabled
-
-  when(if_id.valid) {
-    debugIfInstr(currentThread) := if_id.instr
-  }
-
-  // Advance round-robin selector
-  when(reset.asBool) {
-    threadSel := 0.U
-  } .otherwise {
-    // Fixed round-robin to preserve per-thread spacing (avoids single-thread hazards).
-    val maxThread = (numThreads - 1).U(log2Ceil(numThreads).W)
-    threadSel := Mux(threadSel === maxThread, 0.U, threadSel + 1.U)
-  }
-
-  // ===================== Instruction Decode (ID) with Forwarding =====================
-  val decodeSignals = RV32IDecode.decodeInstr(if_id.instr)
-  val rs1 = if_id.rs1
-  val rs2 = if_id.rs2
-
-  // ID: capture decode outputs into ID/EX register (thread-tagged)
-  id_ex.pc := if_id.pc
-  id_ex.instr := if_id.instr
-  id_ex.threadId := if_id.threadId
-  id_ex.valid := if_id.valid && !flushThread(if_id.threadId)
-  id_ex.rs1 := rs1
-  id_ex.rs2 := rs2
-  id_ex.rd := if_id.rd
-  id_ex.imm := decodeSignals.imm
-  id_ex.aluOp := decodeSignals.aluOp
-  id_ex.isALU := decodeSignals.isALU
-  id_ex.isLoad := decodeSignals.isLoad
-  id_ex.isStore := decodeSignals.isStore
-  id_ex.isBranch := decodeSignals.isBranch
-  id_ex.isJAL := decodeSignals.isJAL
-  id_ex.isJALR := decodeSignals.isJALR
-  id_ex.isLUI := decodeSignals.isLUI
-  id_ex.isAUIPC := decodeSignals.isAUIPC
-  // If this thread is disabled, squash the decode output
-  when(!io.threadEnable(if_id.threadId)) {
-    id_ex.valid := false.B
-  }
-
-  // Register file reads are tagged by threadId
-  regFile.io.readThreadID := if_id.threadId
-  regFile.io.readAddrs(0) := rs1
-  regFile.io.readAddrs(1) := rs2
-
-  val rs1Raw = regFile.io.readData(0)
-  val rs2Raw = regFile.io.readData(1)
-
-  // Simple forwarding when the same thread is in later stages
-  val rs1Fwd = WireDefault(rs1Raw)
-  val rs2Fwd = WireDefault(rs2Raw)
-
-  when(ex_mem.valid && ex_mem.threadId === if_id.threadId && ex_mem.rd =/= 0.U && ex_mem.rd === rs1) {
-    rs1Fwd := Mux(ex_mem.isLoad, memLoadData, ex_mem.aluResult)
-  }
-
-  when(ex_mem.valid && ex_mem.threadId === if_id.threadId && ex_mem.rd =/= 0.U && ex_mem.rd === rs2) {
-    rs2Fwd := Mux(ex_mem.isLoad, memLoadData, ex_mem.aluResult)
-  }
-
-  id_ex.rs1Data := rs1Fwd
-  id_ex.rs2Data := rs2Fwd
-
-  when(id_ex.valid) {
-    debugIdRs1(id_ex.threadId) := rs1Fwd
-    debugIdRs2(id_ex.threadId) := rs2Fwd
-  }
-
-  // ===================== Execute (EX) Stage =====================
-  val alu = Module(new ALU32)
-  val instr = id_ex.instr
-  val opcode = instr(6, 0)
-  val funct3 = instr(14, 12)
-  val funct7 = instr(31, 25)
-  val operandA = WireDefault(id_ex.rs1Data)
-  val operandB = WireDefault(id_ex.rs2Data)
-
-  when(opcode === RV32IDecode.OP_I || opcode === RV32IDecode.LOAD || opcode === RV32IDecode.STORE || opcode === RV32IDecode.JALR) {
-    operandB := id_ex.imm
-  }
-  when(opcode === RV32IDecode.LOAD || opcode === RV32IDecode.STORE || opcode === RV32IDecode.JALR) {
-    operandA := id_ex.rs1Data
-  }
-  when(opcode === RV32IDecode.LUI) {
-    operandA := 0.U
-    operandB := id_ex.imm
-  }
-  when(opcode === RV32IDecode.AUIPC) {
-    operandA := id_ex.pc
-    operandB := id_ex.imm
-  }
-
-  alu.io.a := operandA
-  alu.io.b := operandB
-  alu.io.opcode := id_ex.aluOp
-
-  // ---------- M Extension (Mul/Div) ----------
-  val isMExt = opcode === RV32IDecode.OP_R && funct7 === "b0000001".U
-  val isMulInstr = isMExt && (funct3 <= "b011".U)
-  val isDivInstr = isMExt && (funct3 >= "b100".U)
-
-  val mulUnit = Module(new Mul32OneCycle)
-  val mulSignedA = WireDefault(true.B)
-  val mulSignedB = WireDefault(true.B)
-  when(funct3 === "b011".U) { mulSignedA := false.B; mulSignedB := false.B } // MULHU
-  .elsewhen(funct3 === "b010".U) { mulSignedA := true.B; mulSignedB := false.B } // MULHSU
-  mulUnit.io.a := id_ex.rs1Data
-  mulUnit.io.b := id_ex.rs2Data
-  mulUnit.io.signedA := mulSignedA
-  mulUnit.io.signedB := mulSignedB
-  val mulResult = MuxLookup(funct3, mulUnit.io.lo)(Seq(
-    "b000".U -> mulUnit.io.lo, // MUL
-    "b001".U -> mulUnit.io.hi, // MULH
-    "b010".U -> mulUnit.io.hi, // MULHSU
-    "b011".U -> mulUnit.io.hi  // MULHU
-  ))
-  val mulProductSink = Wire(UInt(64.W))
-  mulProductSink := mulUnit.io.product
-  dontTouch(mulProductSink)
-
-  // Divider wiring
-  divider.io.start := divLaunch
-  divider.io.signed := Mux(divActive, divFunct3 === "b100".U || divFunct3 === "b110".U,
-    isDivInstr && (funct3 === "b100".U || funct3 === "b110".U))
-  divider.io.dividend := Mux(divActive, divDividend, id_ex.rs1Data)
-  divider.io.divisor := Mux(divActive, divDivisor, id_ex.rs2Data)
-  val divBusySink = Wire(UInt(1.W))
-  val divDivideByZeroSink = Wire(UInt(1.W))
-  divBusySink := divider.io.busy
-  divDivideByZeroSink := divider.io.divideByZero
-  dontTouch(divBusySink)
-  dontTouch(divDivideByZeroSink)
-
-  when(!divActive && id_ex.valid && isDivInstr) {
-    divLaunch := true.B
-    divActive := true.B
-    divThread := id_ex.threadId
-    divRd := id_ex.rd
-    divFunct3 := funct3
-    divDividend := id_ex.rs1Data
-    divDivisor := id_ex.rs2Data
-  }.elsewhen(divActive && divider.io.done) {
-    divActive := false.B
-    divDoneSticky := true.B
-    divResult := Mux(divFunct3 === "b100".U || divFunct3 === "b101".U,
-      divider.io.quotient,
-      divider.io.remainder)
-  }
-
-  ex_mem.aluResult := Mux(isMulInstr, mulResult, alu.io.result)
-  ex_mem.instr := id_ex.instr
-  ex_mem.threadId := id_ex.threadId
-  ex_mem.rd := id_ex.rd
-  ex_mem.isALU := id_ex.isALU || isMulInstr
-  ex_mem.isLoad := id_ex.isLoad
-  ex_mem.isStore := id_ex.isStore
-  ex_mem.isBranch := id_ex.isBranch
-  ex_mem.isJAL := id_ex.isJAL
-  ex_mem.isJALR := id_ex.isJALR
-  ex_mem.isLUI := id_ex.isLUI
-  ex_mem.isAUIPC := id_ex.isAUIPC
-  ex_mem.valid := id_ex.valid && !isDivInstr
-  ex_mem.rs1Data := id_ex.rs1Data
-  ex_mem.rs2Data := id_ex.rs2Data
-  ex_mem.pc := id_ex.pc
-  ex_mem.imm := id_ex.imm
-
-  // ===================== Memory (MEM) Stage =====================
-  val loadUnit = Module(new LoadUnit)
-  loadUnit.io.addr := ex_mem.aluResult
-  loadUnit.io.dataIn := io.dataMemResp
-  loadUnit.io.funct3 := ex_mem.instr(14, 12)
-  memLoadData := loadUnit.io.dataOut
-
-  val storeUnit = Module(new StoreUnit)
-  storeUnit.io.addr := ex_mem.aluResult
-  storeUnit.io.data := ex_mem.rs2Data
-  storeUnit.io.storeType := ex_mem.instr(14, 12)
-
-  // Register load data into the ex_mem bundle so forwarding is registered-safe
-  ex_mem.memRdata := memLoadData
-
-  when(ex_mem.valid) {
-    debugExAlu(ex_mem.threadId) := ex_mem.aluResult
-  }
-
-  // Shared memory interface driven by the single current MEM stage
-  val memStoreActive = ex_mem.valid && ex_mem.isStore && io.threadEnable(ex_mem.threadId) && !storeUnit.io.misaligned
-  val addrBase = Cat(ex_mem.aluResult(31, 2), 0.U(2.W))
-  io.memAddr := addrBase
-  io.memWrite := Mux(memStoreActive, storeUnit.io.memWrite, 0.U)
-  io.memMask := Mux(memStoreActive, storeUnit.io.mask, 0.U)
-  val memLoadActive = ex_mem.valid && ex_mem.isLoad && io.threadEnable(ex_mem.threadId)
-  io.memValid := memLoadActive || memStoreActive
-  io.memMisaligned := ex_mem.valid && io.threadEnable(ex_mem.threadId) && storeUnit.io.misaligned
-
-  // MEM+WB combined in `ex_mem` stage; expose load result for debug when valid
-  when(ex_mem.valid) {
-    debugMemLoad(ex_mem.threadId) := memLoadData
-  }
-
-  // ===================== MEM+WB (combined) Stage =====================
-  val pcPlus4 = ex_mem.pc + 4.U
-  val auipcValue = ex_mem.pc + ex_mem.imm
-  val wbData = Wire(UInt(32.W))
-  wbData := ex_mem.aluResult
-  when(ex_mem.isLoad) {
-    wbData := memLoadData
-  }.elsewhen(ex_mem.isLUI) {
-    wbData := ex_mem.imm
-  }.elsewhen(ex_mem.isAUIPC) {
-    wbData := auipcValue
-  }.elsewhen(ex_mem.isJAL || ex_mem.isJALR) {
-    wbData := pcPlus4
-  }
-
-  val writeEnable = ex_mem.valid && ex_mem.rd =/= 0.U &&
-    io.threadEnable(ex_mem.threadId) &&
-    (ex_mem.isALU || ex_mem.isLoad || ex_mem.isLUI ||
-      ex_mem.isAUIPC || ex_mem.isJAL || ex_mem.isJALR)
-
-  val divWrite = divDoneSticky && divRd =/= 0.U && io.threadEnable(divThread) && !writeEnable
-
-  // Safe arbitration: point the regfile write-thread to whichever write wins this cycle.
-  regFile.io.writeThreadID := Mux(divWrite, divThread, ex_mem.threadId)
-  regFile.io.wens(0) := writeEnable || divWrite
-  regFile.io.writeAddrs(0) := Mux(divWrite, divRd, Mux(writeEnable, ex_mem.rd, 0.U))
-  regFile.io.writeData(0) := Mux(divWrite, divResult, wbData)
-
-  when(divWrite) { divDoneSticky := false.B }
-
-  // Writes and reads are tagged with the WB thread ID (now ex_mem)
-  val wbThread = ex_mem.threadId
-
-  // ===================== PC Update & Control =====================
-  val rs1Val = ex_mem.rs1Data
-  val rs2Val = ex_mem.rs2Data
-
-  // Early branch/JAL/JALR resolution at EX to avoid retiring fall-through work
-  val branchCondEx = WireDefault(false.B)
-  switch(id_ex.instr(14, 12)) {
-    is("b000".U) { branchCondEx := id_ex.rs1Data === id_ex.rs2Data }
-    is("b001".U) { branchCondEx := id_ex.rs1Data =/= id_ex.rs2Data }
-    is("b100".U) { branchCondEx := id_ex.rs1Data.asSInt < id_ex.rs2Data.asSInt }
-    is("b101".U) { branchCondEx := !(id_ex.rs1Data.asSInt < id_ex.rs2Data.asSInt) }
-    is("b110".U) { branchCondEx := id_ex.rs1Data < id_ex.rs2Data }
-    is("b111".U) { branchCondEx := !(id_ex.rs1Data < id_ex.rs2Data) }
-  }
-  val branchTakenEx = id_ex.valid && id_ex.isBranch && branchCondEx
-  val branchTargetEx = (id_ex.pc.asSInt + (id_ex.imm.asSInt << 1)).asUInt
-  val jalTakenEx = id_ex.valid && id_ex.isJAL
-  val jalrTakenEx = id_ex.valid && id_ex.isJALR
-  val jalTargetEx = (id_ex.pc.asSInt + id_ex.imm.asSInt).asUInt
-  val jalrTargetEx = ((id_ex.rs1Data.asSInt + id_ex.imm.asSInt).asUInt & ~1.U(32.W))
-
-  // Drive debug/control visibility outputs
-  io.ctrlTaken := branchTakenEx || jalTakenEx || jalrTakenEx
-  io.ctrlThread := id_ex.threadId
-  io.ctrlFromPC := id_ex.pc
-  io.ctrlTarget := Mux(branchTakenEx, branchTargetEx, Mux(jalTakenEx, jalTargetEx, jalrTargetEx))
-  io.ctrlIsBranch := branchTakenEx
-  io.ctrlIsJal := jalTakenEx
-  io.ctrlIsJalr := jalrTakenEx
-
-  when(branchTakenEx && io.threadEnable(id_ex.threadId)) {
-    pcRegs(id_ex.threadId) := branchTargetEx
-    flushThread(id_ex.threadId) := true.B
-  } .elsewhen(jalTakenEx && io.threadEnable(id_ex.threadId)) {
-    pcRegs(id_ex.threadId) := jalTargetEx
-    flushThread(id_ex.threadId) := true.B
-  } .elsewhen(jalrTakenEx && io.threadEnable(id_ex.threadId)) {
-    pcRegs(id_ex.threadId) := jalrTargetEx
-    flushThread(id_ex.threadId) := true.B
-  } .elsewhen(ex_mem.valid) {
-    flushThread(wbThread) := false.B
-  }
-
-  // Flush younger in-flight instructions from the taken-control-transfer thread so fall-through work is discarded.
-  when((branchTakenEx || jalTakenEx || jalrTakenEx) && io.threadEnable(id_ex.threadId)) {
-    when(if_id.threadId === id_ex.threadId) { if_id.valid := false.B }
-  }
-
-  // Default sequential advance for the currently fetched thread unless a control transfer just wrote it.
-  when(!reset.asBool) {
-    when(!flushThisThread && threadEnabled) {
-      when(!(branchTakenEx && id_ex.threadId === currentThread) &&
-           !(jalTakenEx && id_ex.threadId === currentThread) &&
-           !(jalrTakenEx && id_ex.threadId === currentThread)) {
-        pcRegs(currentThread) := currentPC + 4.U
-      }
-      flushThread(currentThread) := false.B
-    } .elsewhen(flushThisThread && threadEnabled) {
-      when(!(branchTakenEx && id_ex.threadId === currentThread) &&
-           !(jalTakenEx && id_ex.threadId === currentThread) &&
-           !(jalrTakenEx && id_ex.threadId === currentThread)) {
-        flushThread(currentThread) := false.B
-      }
-    }
-  }
-
-  // Hold disabled threads in place and clear any pending flush for them.
-  for (t <- 0 until numThreads) {
-    when(!io.threadEnable(t)) {
-      pcRegs(t) := pcRegs(t)
-      flushThread(t) := false.B
-    }
-  }
-
-  // Commit next-state values
-  when(reset.asBool) {
-    pcRegs := pcResetVec
-    flushThread := flushResetVec
-  }
-
-  // ===================== Expose Pipeline State =====================
-  io.if_pc := pcRegs
-  io.if_instr := debugIfInstr
-  io.id_rs1Data := debugIdRs1
-  io.id_rs2Data := debugIdRs2
-  io.ex_aluResult := debugExAlu
-  io.mem_loadData := debugMemLoad
-  // Default debug outputs when no control transfer
-  when(!io.ctrlTaken) {
-    io.ctrlThread := 0.U
-    io.ctrlFromPC := 0.U
-    io.ctrlTarget := 0.U
-  }
-}
