+++ rtl/generators/GenerateHierarchicalRTL.scala
@@ -21,7 +21,7 @@ import Decoders.RV32IDecodeModule
 import LoadUnit.LoadUnit
 import RegFiles.{RegFileMT2R1WMem, RegFileMT2R1WVec}
 import StoreUnit.StoreUnit
-import TetraNyte.{TetraNyteRV32ICore, TetraNyteRV32ICoreMRF}
+import TetraNyte.TetraNyteRV32ICore
 import ZeroNyte.{ZeroNyteRV32ICore, ZeroNyteRV32ICoreWithCache, ZeroNyteRV32IZmmulCore, ZeroNyteRV32IMCore}
 import OctoNyte.OctoNyteRV32ICore
 
@@ -315,10 +315,8 @@ Environment Variables:
       case "rv32i" =>
         // Generate all building blocks plus the threaded core itself
         val libraryBlocks = getRV32ILibraryModules("TetraNyte")
-        libraryBlocks ++ Seq(
-          ModuleSpec(() => new TetraNyteRV32ICore, "TetraNyteRV32ICore", "Four-thread barrel-threaded RV32I core", "TetraNyte", "rv32i"),
-          ModuleSpec(() => new TetraNyteRV32ICoreMRF, "TetraNyteRV32ICoreMRF", "Four-thread barrel-threaded RV32I core with MRF", "TetraNyte", "rv32i")
-        )
+        libraryBlocks :+
+          ModuleSpec(() => new TetraNyteRV32ICore, "TetraNyteRV32ICore", "Four-thread barrel-threaded RV32I core", "TetraNyte", "rv32i")
       case _ => Seq.empty
     }
   }
